+++ src/python/pants/backend/docker/dependencies.py
@@ -2,11 +2,15 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 import re
-from typing import Generator, Optional
+from dataclasses import dataclass
+from typing import Generator, Optional, Tuple
+
+from pants.backend.docker.dockerfile import Dockerfile
 
 
-class DockerfileDependenciesMixin:
-    """Mixin class for the Dockerfile class."""
+@dataclass
+class DockerfileDependencies:
+    dockerfile: Dockerfile
 
     _pex_target_regexp = re.compile(
         r"""
@@ -15,9 +19,13 @@ class DockerfileDependenciesMixin:
         re.VERBOSE,
     )
 
-    def putative_target_addresses(self) -> Generator[str, None, None]:
+    def putative_target_addresses(self) -> Tuple[str, ...]:
         """Inspect the Dockerfile for potential dependencies."""
-        for copy in self.copy:
+        return tuple(self._putative_target_addresses())
+
+    def _putative_target_addresses(self) -> Generator[str, None, None]:
+        """Inspect the Dockerfile for potential dependencies."""
+        for copy in self.dockerfile.copy:
             for src in copy.src:
                 addr = self.as_address(src)
                 if addr:
